refacto in specs
